agents/requirements.txt
uagents>=0.22.0
uagents-core>=0.1.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
httpx>=0.26.0
numpy>=1.24.0
scikit-learn>=1.3.0
openai>=1.0.0
hyperon>=0.1.12
pillow>=10.0.0
